reorganize and add test
