+++ preliz/predictive/__init__.py
@@ -1,0 +1,1 @@
+from .ppa import ppa
