+++ preliz/unidimensional/__init__.py
@@ -1,0 +1,4 @@
+from .maxent import maxent
+from .mle import mle
+from .quartile import quartile
+from .roulette import roulette
